Solve style consistency issues for messages
Uniformise error messages in src/parser.js with remaining messages in
the project. Uniformise style in help message, fix a typo and update
the usage of 'noindex' to also include the 'classes' option. Add 'XML'
to the usage string to be more informative. Update README.md to reflect
the changes.